modularized pubsub; Flooding is just a routing method.
