+++ notify.go
@@ -17,7 +17,7 @@ func (p *PubSubNotif) ClosedStream(n inet.Network, s inet.Stream) {
 
 func (p *PubSubNotif) Connected(n inet.Network, c inet.Conn) {
 	go func() {
-		s, err := p.host.NewStream(p.ctx, c.RemotePeer(), ID)
+		s, err := p.host.NewStream(p.ctx, c.RemotePeer(), p.rt.Protocols()...)
 		if err != nil {
 			log.Warning("opening new stream to peer: ", err, c.LocalPeer(), c.RemotePeer())
 			return
